+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/imports/converters/TensorFlowMapper.java
@@ -1,5 +1,6 @@
 package org.nd4j.imports.converters;
 
+import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.nd4j.autodiff.execution.Node;
 import org.nd4j.graph.intermediate.TGraph;
@@ -18,6 +19,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
+@Slf4j
 public class TensorFlowMapper implements NodeMapper<NodeDef> {
     private static final TensorFlowMapper INSTANCE = new TensorFlowMapper();
     private Map<String, ExternalNode<NodeDef>> nodeConverters = new HashMap<>();
@@ -42,8 +44,10 @@ public class TensorFlowMapper implements NodeMapper<NodeDef> {
                 val name = node.opName();
                 if (nodeConverters.containsKey(name)) {
                     throw new ND4JIllegalStateException("OpName duplicate found: " + name);
-                } else
+                } else {
+                    log.info("Adding converter for [" + name + "]");
                     nodeConverters.put(name, node);
+                }
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }
@@ -56,7 +60,15 @@ public class TensorFlowMapper implements NodeMapper<NodeDef> {
 
     @Override
     public TNode asIntermediate(NodeDef node, TGraph graph) {
-        return nodeConverters.get(node.getOp()).asIntermediateRepresentation(node, graph);
+        // first we try to use special converters
+        val converter = nodeConverters.get(node.getOp().toLowerCase());
+        if (converter != null)
+            return converter.asIntermediateRepresentation(node, graph);
+        else {
+            val defaultConverter = nodeConverters.get("GenericOpConverter");
+
+            return defaultConverter.asIntermediateRepresentation(node, graph);
+        }
     }
 
     public Set<String> knownOps() {
